+++ json-forms-compose/src/commonMain/kotlin/com/copperleaf/forms/compose/controls/testers.kt
@@ -1,6 +1,7 @@
 package com.copperleaf.forms.compose.controls
 
 import com.copperleaf.forms.core.ControlType
+import com.copperleaf.forms.core.StringControl
 import com.copperleaf.forms.core.ui.UiElement
 import com.copperleaf.json.values.boolean
 import com.copperleaf.json.values.objectAt
@@ -26,4 +27,23 @@ public fun UiElement.Control.optionFieldIs(name: String, value: String): Boolean
 
 public fun UiElement.Control.hasSchemaProperty(name: String): Boolean {
     return schemaConfig.jsonObject.containsKey(name)
+}
+
+public fun UiElement.Control.hasArrayItemType(type: StringControl): Boolean {
+    return schemaConfig
+        .optional { objectAt("items").string("type") } == type.type
+}
+
+public fun UiElement.Control.hasArrayItemProperty(name: String): Boolean {
+    return schemaConfig
+        .optional { objectAt("items").containsKey(name) } == true
+}
+
+public fun UiElement.Control.schemaPropertyIs(name: String, value: String): Boolean {
+    return schemaConfig.optional { string(name) } == value
+}
+
+public fun UiElement.Control.schemaPropertyIsEnabled(name: String): Boolean {
+    return schemaConfig
+        .optional { boolean(name) } == true
 }